pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adaptive-fake-news-detector"
version = "1.0.0"
description = "A comprehensive FastAPI backend for real-time news data collection and fake news detection"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.8"

dependencies = [
    # Core FastAPI dependencies
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    
    # HTTP client and web scraping
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    
    # RSS feed parsing
    "feedparser>=6.0.10",
    
    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dateutil>=2.8.2",
    
    # Configuration and environment
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    
    # File handling
    "aiofiles>=23.0.0",
    
    # Background tasks (optional)
    "celery>=5.3.0",
    "redis>=5.0.0",
    
    # Database (for future use)
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",  # for testing
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

ml = [
    "scikit-learn>=1.3.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "shap>=0.42.0",
]

monitoring = [
    "prometheus-client>=0.17.0",
    "structlog>=23.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/adaptive-fake-news-detector"
Repository = "https://github.com/yourusername/adaptive-fake-news-detector"
Issues = "https://github.com/yourusername/adaptive-fake-news-detector/issues"

[project.scripts]
fake-news-api = "app.main:main"

[tool.ruff]
target-version = "py38"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "feedparser.*",
    "bs4.*",
]
ignore_missing_imports = true 